fix wrong usages of cJSON::Create/cJSON::CreateArray which cause memleak
When we create a JSONValue with flag false, which means a non-root
JSONValue. When the JSONValue destructes, it will not free the opaque
cJSON object. So when this JSONValue is put into another JSONValue,
the opaque cJSON object is deep copied, and the origin cJSON object
will not be freed, which finally causes the memleak.

Signed-off-by: Far <yesiyuan2@huawei.com>
Change-Id: I6ffdc3a80d6b1551e733e92b6329af0496048973